+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 6.1.2 - 2023-07-23
+
+### Changed
+* Update FCS to 'Add trivia information to SynConst.Measure', commit 7b5e12842d673b7daa467e0091378bf4acc95e4f
+
+### Fixed
+* Left out lhs in SynMeasure.Divide should not be restored as SynMeasure.One. [#2926](https://github.com/fsprojects/fantomas/issues/2926)
+* Block comments in measure are lost or restored twice and in wrong place. [#2927](https://github.com/fsprojects/fantomas/issues/2927)
+
 ## 6.1.1 - 2023-06-29
 
 ### Changed